requirements.txt
requests ~= 2.25.1
PyYAML ~= 5.3.1
python-dateutil ~= 2.8.1